updated changes in password protection
