+++ NeuroDrive/serializers/__init__.py
@@ -1,8 +1,10 @@
 from NeuroDrive.serializers.file import FileSerializer
+from NeuroDrive.serializers.fileAccess import FileAccessSerializer
 from NeuroDrive.serializers.directory import DirectorySerializer
 
 
 __all__ = [
     "FileSerializer",
     "DirectorySerializer",
+    "FileAccessSerializer"
 ]
